cgroup: Move cgroup handling to separate files in libutil.
