+++ src/libutil/Makefile.am
@@ -1,12 +1,13 @@
 pkglib_LTLIBRARIES = libutil.la
 
 libutil_la_SOURCES = util.cc hash.cc serialise.cc \
-  archive.cc xml-writer.cc affinity.cc
+  archive.cc xml-writer.cc affinity.cc cgroups.cc
 
 libutil_la_LIBADD = ../boost/format/libformat.la
 
 pkginclude_HEADERS = util.hh hash.hh serialise.hh \
-  archive.hh xml-writer.hh types.hh affinity.hh
+  archive.hh xml-writer.hh types.hh affinity.hh \
+  cgroups.hh
 
 if !HAVE_OPENSSL
 libutil_la_SOURCES += \
