+++ src/libutil/util.hh
@@ -353,7 +353,4 @@ string decodeOctalEscaped(const string & s);
    ignore the exception. */
 void ignoreException();
 
-Cgroups getCgroups(long pid = -1);
-void joinCgroups(const Cgroups &);
-
 }
